// src/interfaces/ILogFilePaths.ts
export interface ILogFilePaths {
  /** The path to the log file for debug level messages. */
  debug: string

  /** The path to the log file for error level messages. */
  error: string

  /** The path to the log file for info level messages. */
  info: string

  /** The path to the log file for success level messages. */
  success: string

  /** The path to the log file for warn level messages. */
  warn: string
}